Changed controller methods so the actions would be done one at a time, instead of everything being processed in return statement.
