+++ controller/src/main/java/com/softserve/borda/controllers/BoardController.java
@@ -1,19 +1,11 @@
 package com.softserve.borda.controllers;
 
 import com.softserve.borda.config.jwt.JwtConvertor;
-import com.softserve.borda.dto.BoardFullDTO;
-import com.softserve.borda.dto.BoardListDTO;
-import com.softserve.borda.dto.CreateBoardDTO;
-import com.softserve.borda.entities.Board;
-import com.softserve.borda.entities.BoardList;
-import com.softserve.borda.entities.BoardRole;
-import com.softserve.borda.entities.UserBoardRelation;
+import com.softserve.borda.dto.*;
+import com.softserve.borda.entities.*;
 import com.softserve.borda.exceptions.CustomEntityNotFoundException;
 import com.softserve.borda.exceptions.CustomFailedToDeleteEntityException;
-import com.softserve.borda.services.BoardListService;
-import com.softserve.borda.services.BoardService;
-import com.softserve.borda.services.UserBoardRelationService;
-import com.softserve.borda.services.UserService;
+import com.softserve.borda.services.*;
 import lombok.AllArgsConstructor;
 import lombok.extern.java.Log;
 import org.modelmapper.ModelMapper;
@@ -43,6 +35,12 @@ public class BoardController {
 
     private final UserBoardRelationService userBoardRelationService;
 
+    private final CommentService commentService;
+
+    private final TagService tagService;
+
+    private final TicketService ticketService;
+
     private final JwtConvertor jwtConvertor;
 
     @GetMapping
@@ -53,6 +51,36 @@ public class BoardController {
     @GetMapping("/{id}")
     public Board getBoard(@PathVariable Long id) {
         return boardService.getBoardById(id);
+    }
+
+    @GetMapping("/{userId}/boards")
+    public ResponseEntity<List<BoardFullDTO>> getBoardsByUserId(@PathVariable Long userId) {
+        try {
+            List<Board> boards = userService.getBoardsByUserId(userId);
+            List<BoardFullDTO> boardDTOs = boards.stream().map(board -> modelMapper.map(board,
+                    BoardFullDTO.class)).collect(Collectors.toList());
+
+            return new ResponseEntity<>(boardDTOs, HttpStatus.OK);
+        } catch (CustomEntityNotFoundException e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @GetMapping("{userId}/boardsByRole/{boardRoleId}")
+    public ResponseEntity<List<BoardFullDTO>> getBoardsByUserIdAndBoardRoleId(
+            @PathVariable Long userId,
+            @PathVariable Long boardRoleId) {
+        try {
+            List<Board> boards = userService.getBoardsByUserIdAndBoardRoleId(userId, boardRoleId);
+            List<BoardFullDTO> boardDTOs = boards.stream().map(board -> modelMapper.map(board,
+                    BoardFullDTO.class)).collect(Collectors.toList());
+
+            return new ResponseEntity<>(boardDTOs, HttpStatus.OK);
+        } catch (CustomEntityNotFoundException e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @PostMapping("createBoard/{userId}")
@@ -71,9 +99,10 @@ public class BoardController {
 
             board.setUserBoardRelations(Collections.singletonList(userBoardRelation));
 
-            return new ResponseEntity<>(modelMapper.map(
-                    boardService.createOrUpdate(board),
-                    BoardFullDTO.class), HttpStatus.CREATED);
+            board = boardService.createOrUpdate(board);
+            BoardFullDTO boardFullDTO = modelMapper.map(board, BoardFullDTO.class);
+
+            return new ResponseEntity<>(boardFullDTO, HttpStatus.CREATED);
         } catch (Exception e) {
             log.severe(e.getMessage());
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
@@ -84,6 +113,7 @@ public class BoardController {
     public ResponseEntity<String> deleteBoard(@PathVariable Long id) {
         try {
             boardService.deleteBoardById(id);
+
             return new ResponseEntity<>("Entity was removed successfully",
                     HttpStatus.OK);
         } catch (CustomFailedToDeleteEntityException e) {
@@ -95,13 +125,14 @@ public class BoardController {
 
     @PutMapping(value = "{id}")
     public ResponseEntity<BoardFullDTO> updateBoard(@PathVariable Long id,
-                                                    @RequestBody BoardFullDTO board) {
+                                                    @RequestBody BoardFullDTO boardFullDTO) {
         try {
-            Board existingBoard = boardService.getBoardById(id);
-            BeanUtils.copyProperties(board, existingBoard);
-            return new ResponseEntity<>(modelMapper.map(
-                    boardService.createOrUpdate(existingBoard),
-                    BoardFullDTO.class), HttpStatus.OK);
+            Board board = boardService.getBoardById(id);
+            BeanUtils.copyProperties(boardFullDTO, board);
+            board = boardService.createOrUpdate(board);
+            boardFullDTO = modelMapper.map(board, BoardFullDTO.class);
+
+            return new ResponseEntity<>(boardFullDTO, HttpStatus.OK);
         } catch (Exception e) {
             log.severe(e.getMessage());
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -111,11 +142,12 @@ public class BoardController {
     @GetMapping("{id}/boardLists")
     public ResponseEntity<List<BoardListDTO>> getAllBoardListsForBoard(@PathVariable Long id) {
         try {
-            return new ResponseEntity<>(
-                    boardService.getAllBoardListsByBoardId(id)
-                            .stream().map(boardList -> modelMapper.map(boardList,
-                            BoardListDTO.class)).collect(Collectors.toList()),
-                    HttpStatus.OK);
+            List<BoardList> boardLists = boardService.getAllBoardListsByBoardId(id);
+            List<BoardListDTO> boardListDTOs = boardLists.stream()
+                    .map(boardList -> modelMapper.map(boardList,
+                            BoardListDTO.class)).collect(Collectors.toList());
+
+            return new ResponseEntity<>(boardListDTOs, HttpStatus.OK);
         } catch (CustomEntityNotFoundException e) {
             log.severe(e.getMessage());
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -129,12 +161,11 @@ public class BoardController {
             BoardList boardList = new BoardList();
             boardList.setName(boardListDTO.getName());
             boardList = boardListService.createOrUpdate(boardList);
-            return new ResponseEntity<>(
-                    modelMapper.map(
-                            boardService.addBoardListToBoard(
-                                    boardService.getBoardById(boardId), boardList),
-                            BoardListDTO.class),
-                    HttpStatus.OK);
+            boardList = boardService.addBoardListToBoard(
+                    boardService.getBoardById(boardId), boardList);
+            boardListDTO = modelMapper.map(boardList, BoardListDTO.class);
+
+            return new ResponseEntity<>(boardListDTO, HttpStatus.OK);
         } catch (Exception e) {
             log.severe(e.getMessage());
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
@@ -147,6 +178,7 @@ public class BoardController {
         try {
             boardService.deleteBoardListFromBoard(boardService.getBoardById(boardId),
                     boardListService.getBoardListById(boardListId));
+
             return new ResponseEntity<>("Entity was removed successfully",
                     HttpStatus.OK);
         } catch (CustomFailedToDeleteEntityException e) {
@@ -155,4 +187,349 @@ public class BoardController {
                     HttpStatus.NOT_FOUND);
         }
     }
+
+    @GetMapping("/{boardListId}")
+    public ResponseEntity<BoardListDTO> getBoardListById(@PathVariable Long boardListId) {
+        try {
+            BoardList boardList = boardListService.getBoardListById(boardListId);
+            BoardListDTO boardListDTO = modelMapper.map(boardList, BoardListDTO.class);
+
+            return new ResponseEntity<>(boardListDTO, HttpStatus.OK);
+        } catch (CustomEntityNotFoundException e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @PutMapping(value = "{boardListId}")
+    public ResponseEntity<BoardListDTO> updateBoardList(@PathVariable Long boardListId,
+                                                        @RequestBody BoardListDTO boardListDTO) {
+        try {
+            BoardList boardList = boardListService.getBoardListById(boardListId);
+            BeanUtils.copyProperties(boardListDTO, boardList);
+            boardList = boardListService.createOrUpdate(boardList);
+            boardListDTO = modelMapper.map(boardList, BoardListDTO.class);
+
+            return new ResponseEntity<>(boardListDTO, HttpStatus.OK);
+        } catch (Exception e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @GetMapping("{boardListId}/tickets")
+    public ResponseEntity<List<TicketDTO>> getAllTicketsForBoardList(@PathVariable Long boardListId) {
+        try {
+            List<Ticket> tickets = boardListService.getAllTicketsByBoardListId(boardListId);
+            List<TicketDTO> ticketDTOs = tickets.stream()
+                    .map(ticket -> modelMapper.map(ticket,
+                            TicketDTO.class)).collect(Collectors.toList());
+
+            return new ResponseEntity<>(ticketDTOs, HttpStatus.OK);
+        } catch (CustomEntityNotFoundException e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @PostMapping("{boardListId}/addTicket")
+    public ResponseEntity<BoardListDTO> createTicketForBoardList(@PathVariable long boardListId,
+                                                                 @RequestBody TicketDTO ticketDTO) {
+        try {
+            Ticket ticket = modelMapper.map(ticketDTO, Ticket.class);
+            ticket = ticketService.createOrUpdate(ticket);
+            BoardList boardList = boardListService.getBoardListById(boardListId);
+            boardList = boardListService.addTicketToBoardList(boardList, ticket);
+            BoardListDTO boardListDTO = modelMapper.map(boardList, BoardListDTO.class);
+
+            return new ResponseEntity<>(boardListDTO, HttpStatus.OK);
+        } catch (Exception e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @DeleteMapping(value = "{boardListId}/tickets/{ticketId}")
+    public ResponseEntity<BoardListDTO> deleteTicketFromBoardList(@PathVariable Long boardListId,
+                                                                  @PathVariable Long ticketId) {
+        try {
+            BoardList boardList = boardListService.getBoardListById(boardListId);
+            Ticket ticket = ticketService.getTicketById(ticketId);
+            boardList = boardListService.deleteTicketFromBoardList(boardList,
+                    ticket);
+            BoardListDTO boardListDTO = modelMapper.map(boardList, BoardListDTO.class);
+
+            return new ResponseEntity<>(boardListDTO, HttpStatus.OK);
+        } catch (CustomFailedToDeleteEntityException e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @PostMapping("/moveFrom/{oldBoardListId}/to/{newBoardListId}/ticket/{ticketId}/")
+    public ResponseEntity<BoardListDTO> moveTicketToAnotherBoardList(@PathVariable Long oldBoardListId,
+                                                                     @PathVariable Long newBoardListId,
+                                                                     @PathVariable Long ticketId) {
+        try {
+            BoardList oldBoardList = boardListService.getBoardListById(oldBoardListId);
+            BoardList newBoardList = boardListService.getBoardListById(newBoardListId);
+            Ticket ticket = ticketService.getTicketById(ticketId);
+            oldBoardList.getTickets().remove(ticket);
+            newBoardList.getTickets().add(ticket);
+            boardListService.createOrUpdate(oldBoardList);
+            newBoardList = boardListService.createOrUpdate(newBoardList);
+            BoardListDTO boardListDTO = modelMapper.map(newBoardList, BoardListDTO.class);
+
+            return new ResponseEntity<>(boardListDTO, HttpStatus.OK);
+        } catch (CustomEntityNotFoundException e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @GetMapping("/tickets/{ticketId}")
+    public ResponseEntity<TicketDTO> getTicketById(@PathVariable Long ticketId) {
+        try {
+            Ticket ticket = ticketService.getTicketById(ticketId);
+            TicketDTO ticketDTO = modelMapper.map(ticket, TicketDTO.class);
+
+            return new ResponseEntity<>(ticketDTO, HttpStatus.OK);
+        } catch (CustomEntityNotFoundException e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @PutMapping(value = "/tickets/{ticketId}")
+    public ResponseEntity<TicketDTO> updateTicket(@PathVariable Long ticketId,
+                                                  @RequestBody TicketDTO ticketDTO) {
+        try {
+            Ticket ticket = ticketService.getTicketById(ticketId);
+            BeanUtils.copyProperties(ticketDTO, ticket);
+            ticket = ticketService.createOrUpdate(ticket);
+            ticketDTO = modelMapper.map(ticket, TicketDTO.class);
+
+            return new ResponseEntity<>(ticketDTO, HttpStatus.OK);
+        } catch (Exception e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @GetMapping("/tickets/{ticketId}/comments")
+    public ResponseEntity<List<CommentDTO>> getCommentsByTicketId(@PathVariable Long ticketId) {
+        try {
+            List<Comment> comments = ticketService.getAllCommentsByTicketId(ticketId);
+            List<CommentDTO> commentDTOs = comments.stream().map(comment ->
+                    modelMapper.map(comment, CommentDTO.class))
+                    .collect(Collectors.toList());
+
+            return new ResponseEntity<>(commentDTOs, HttpStatus.OK);
+        } catch (CustomEntityNotFoundException e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @GetMapping("/tickets/{ticketId}/tags")
+    public ResponseEntity<List<TagDTO>> getTagsByTicketId(@PathVariable Long ticketId) {
+        try {
+            List<Tag> tags = ticketService.getAllTagsByTicketId(ticketId);
+            List<TagDTO> tagDTOs = tags.stream().map(tag ->
+                    modelMapper.map(tag, TagDTO.class))
+                    .collect(Collectors.toList());
+
+            return new ResponseEntity<>(tagDTOs, HttpStatus.OK);
+        } catch (CustomEntityNotFoundException e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @GetMapping("/tickets/{ticketId}/members")
+    public ResponseEntity<List<UserSimpleDTO>> getMembersByTicketId(@PathVariable Long ticketId) {
+        try {
+            List<User> users = ticketService.getAllMembersByTicketId(ticketId);
+            List<UserSimpleDTO> userDTOs = users.stream().map(user ->
+                    modelMapper.map(user, UserSimpleDTO.class))
+                    .collect(Collectors.toList());
+
+            return new ResponseEntity<>(userDTOs, HttpStatus.OK);
+        } catch (CustomEntityNotFoundException e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @GetMapping("/comments/{commentId}")
+    public ResponseEntity<CommentDTO> getCommentByCommentId(@PathVariable Long commentId) {
+        try {
+            Comment comment = commentService.getCommentById(commentId);
+            CommentDTO commentDTO = modelMapper.map(comment, CommentDTO.class);
+
+            return new ResponseEntity<>(commentDTO, HttpStatus.OK);
+        } catch (CustomEntityNotFoundException e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @PostMapping("/tickets/{ticketId}/addComment")
+    public ResponseEntity<TicketDTO> addCommentToTicketAndUser(@PathVariable Long ticketId,
+                                                               @RequestBody CommentDTO commentDTO) {
+        try {
+            Comment comment = new Comment();
+            comment.setText(commentDTO.getText());
+            comment.setUser(userService.getUserById(commentDTO.getUserId()));
+            comment = commentService.createOrUpdate(comment);
+            userService.addCommentToUser(comment.getUser().getId(), comment);
+            Ticket ticket = ticketService.addCommentToTicket(ticketId, comment);
+            TicketDTO ticketDTO = modelMapper.map(ticket, TicketDTO.class);
+
+            return new ResponseEntity<>(ticketDTO, HttpStatus.OK);
+        } catch (Exception e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @PutMapping("/comments/{commentId}")
+    public ResponseEntity<CommentDTO> updateCommentByCommentId(@PathVariable Long commentId,
+                                                               @RequestBody CommentDTO commentDTO) {
+        try {
+            Comment comment = commentService.getCommentById(commentId);
+            BeanUtils.copyProperties(commentDTO, comment);
+            comment = commentService.createOrUpdate(comment);
+            commentDTO = modelMapper.map(comment, CommentDTO.class);
+
+            return new ResponseEntity<>(commentDTO, HttpStatus.OK);
+        } catch (CustomEntityNotFoundException e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @DeleteMapping("/tickets/{ticketId}/deleteComment/{commentId}")
+    public ResponseEntity<TicketDTO> deleteCommentFromTicketAndUser(@PathVariable Long ticketId,
+                                                                    @PathVariable Long commentId) {
+        try {
+            Comment comment = commentService.getCommentById(commentId);
+            ticketService.deleteCommentFromTicket(ticketId, comment);
+            userService.deleteCommentFromUser(comment.getUser().getId(), comment);
+            commentService.deleteCommentById(commentId);
+            Ticket ticket = ticketService.deleteCommentFromTicket(ticketId, comment);
+            TicketDTO ticketDTO = modelMapper.map(ticket, TicketDTO.class);
+
+            return new ResponseEntity<>(ticketDTO, HttpStatus.OK);
+        } catch (CustomFailedToDeleteEntityException e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @GetMapping("/tags/{tagId}")
+    public ResponseEntity<TagDTO> getTagByTagId(@PathVariable Long tagId) {
+        try {
+            Tag tag = tagService.getTagById(tagId);
+            TagDTO tagDTO = modelMapper.map(tag, TagDTO.class);
+
+            return new ResponseEntity<>(tagDTO, HttpStatus.OK);
+        } catch (CustomEntityNotFoundException e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @PostMapping("/tickets/{ticketId}/addTag")
+    public ResponseEntity<TicketDTO> addTagToTicket(@PathVariable Long ticketId,
+                                                    @RequestBody TagDTO tagDTO) {
+        try {
+            Tag tag = modelMapper.map(tagDTO, Tag.class);
+            tag = tagService.createOrUpdate(tag);
+            Ticket ticket = ticketService.addTagToTicket(ticketId, tag);
+            TicketDTO ticketDTO = modelMapper.map(ticket, TicketDTO.class);
+
+            return new ResponseEntity<>(ticketDTO, HttpStatus.OK);
+        } catch (Exception e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @GetMapping("/tagList")
+    public ResponseEntity<List<TagDTO>> listOfTags() {
+        try {
+            List<Tag> tags = tagService.getAll();
+            List<TagDTO> tagDTOs = tags.stream()
+                    .map(user -> modelMapper.map(user, TagDTO.class))
+                    .collect(Collectors.toList());
+
+            return new ResponseEntity<>(tagDTOs, HttpStatus.OK);
+        } catch (Exception e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @PutMapping("/tags/{tagId}")
+    public ResponseEntity<TagDTO> updateTagByTagId(@PathVariable Long tagId,
+                                                   @RequestBody TagDTO tagDTO) {
+        try {
+            Tag tag = tagService.getTagById(tagId);
+            BeanUtils.copyProperties(tagDTO, tag);
+            tag = tagService.createOrUpdate(tag);
+            tagDTO = modelMapper.map(tag, TagDTO.class);
+
+            return new ResponseEntity<>(tagDTO, HttpStatus.OK);
+        } catch (CustomEntityNotFoundException e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @DeleteMapping("/tickets/{ticketId}/deleteTag/{tagId}")
+    public ResponseEntity<TicketDTO> deleteTagFromTicket(@PathVariable Long ticketId,
+                                                         @PathVariable Long tagId) {
+        try {
+            Tag tag = tagService.getTagById(tagId);
+            Ticket ticket = ticketService.deleteTagFromTicket(ticketId, tag);
+            TicketDTO ticketDTO = modelMapper.map(ticket, TicketDTO.class);
+
+            return new ResponseEntity<>(ticketDTO, HttpStatus.OK);
+        } catch (CustomFailedToDeleteEntityException e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @PostMapping("/tickets/{ticketId}/addMember/{userId}")
+    public ResponseEntity<TicketDTO> addUserToTicket(@PathVariable Long ticketId,
+                                                     @PathVariable Long userId) {
+        try {
+            User user = userService.getUserById(userId);
+            Ticket ticket = ticketService.addMemberToTicket(ticketId, user);
+            TicketDTO ticketDTO = modelMapper.map(ticket, TicketDTO.class);
+
+            return new ResponseEntity<>(ticketDTO, HttpStatus.OK);
+        } catch (Exception e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @DeleteMapping("/tickets/{ticketId}/deleteMember/{userId}")
+    public ResponseEntity<TicketDTO> deleteUserFromTicket(@PathVariable Long ticketId,
+                                                          @PathVariable Long userId) {
+        try {
+            User user = userService.getUserById(userId);
+            Ticket ticket = ticketService.deleteMemberFromTicket(ticketId, user);
+            TicketDTO ticketDTO = modelMapper.map(ticket, TicketDTO.class);
+
+            return new ResponseEntity<>(ticketDTO, HttpStatus.OK);
+        } catch (CustomFailedToDeleteEntityException e) {
+            log.severe(e.getMessage());
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
 }
