+++ controller/src/main/java/com/softserve/borda/controllers/UserController.java
@@ -1,7 +1,9 @@
 package com.softserve.borda.controllers;
 
 import com.softserve.borda.config.jwt.JwtConvertor;
-import com.softserve.borda.dto.*;
+import com.softserve.borda.dto.CreateUserDTO;
+import com.softserve.borda.dto.UserFullDTO;
+import com.softserve.borda.dto.UserSimpleDTO;
 import com.softserve.borda.entities.User;
 import com.softserve.borda.exceptions.CustomEntityNotFoundException;
 import com.softserve.borda.exceptions.CustomFailedToDeleteEntityException;
@@ -33,9 +35,12 @@ public class UserController {
     @GetMapping
     public ResponseEntity<List<UserSimpleDTO>> getAllUsers() {
         try {
-            return new ResponseEntity<>(userService.getAll().stream()
+            List<User> users = userService.getAll();
+            List<UserSimpleDTO> userDTOs = users.stream()
                     .map(user -> modelMapper.map(user, UserSimpleDTO.class))
-                    .collect(Collectors.toList()), HttpStatus.OK);
+                    .collect(Collectors.toList());
+
+            return new ResponseEntity<>(userDTOs, HttpStatus.OK);
         } catch (Exception e) {
             log.severe(e.getMessage());
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
@@ -45,10 +50,10 @@ public class UserController {
     @GetMapping("getById/{id}")
     public ResponseEntity<UserSimpleDTO> getUserById(@PathVariable Long id) {
         try {
-            return new ResponseEntity<>(modelMapper.map(
-                    userService.getUserById(id)
-                    ,
-                    UserSimpleDTO.class), HttpStatus.OK);
+            User user = userService.getUserById(id);
+            UserSimpleDTO userSimpleDTO = modelMapper.map(user, UserSimpleDTO.class);
+
+            return new ResponseEntity<>(userSimpleDTO, HttpStatus.OK);
         } catch (CustomEntityNotFoundException e) {
             log.severe(e.getMessage());
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -58,9 +63,10 @@ public class UserController {
     @GetMapping("{id}")
     public ResponseEntity<UserSimpleDTO> getUserById(@PathVariable String id) {
         try {
-            return new ResponseEntity<>(modelMapper.map(
-                    jwtConvertor.getUserByJWT(id),
-                    UserSimpleDTO.class), HttpStatus.OK);
+            User user = jwtConvertor.getUserByJWT(id);
+            UserSimpleDTO userSimpleDTO = modelMapper.map(user, UserSimpleDTO.class);
+
+            return new ResponseEntity<>(userSimpleDTO, HttpStatus.OK);
         } catch (CustomEntityNotFoundException e) {
             log.severe(e.getMessage());
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -71,9 +77,10 @@ public class UserController {
     public ResponseEntity<UserSimpleDTO> createUser(@RequestBody CreateUserDTO userDTO) {
         try {
             User user = modelMapper.map(userDTO, User.class);
-            return new ResponseEntity<>(modelMapper.map(
-                    userService.createOrUpdate(user),
-                    UserSimpleDTO.class), HttpStatus.CREATED);
+            user = userService.createOrUpdate(user);
+            UserSimpleDTO userSimpleDTO = modelMapper.map(user, UserSimpleDTO.class);
+
+            return new ResponseEntity<>(userSimpleDTO, HttpStatus.CREATED);
         } catch (Exception e) {
             log.severe(e.getMessage());
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
@@ -84,6 +91,7 @@ public class UserController {
     public ResponseEntity<String> deleteUserById(@PathVariable Long id) {
         try {
             userService.deleteUserById(id);
+
             return new ResponseEntity<>("Entity was removed successfully",
                     HttpStatus.OK);
         } catch (CustomFailedToDeleteEntityException e) {
@@ -95,67 +103,15 @@ public class UserController {
 
     @PutMapping(value = "{id}")
     public ResponseEntity<UserSimpleDTO> updateUser(@PathVariable Long id,
-                                                    @RequestBody UserFullDTO user) {
+                                                    @RequestBody UserFullDTO userFullDTO) {
         try {
-            User existingUser = userService.getUserById(id);
-            BeanUtils.copyProperties(user, existingUser);
-            return new ResponseEntity<>(modelMapper.map(
-                    userService.createOrUpdate(existingUser),
-                    UserSimpleDTO.class), HttpStatus.OK);
+            User user = userService.getUserById(id);
+            BeanUtils.copyProperties(userFullDTO, user);
+            user = userService.createOrUpdate(user);
+            UserSimpleDTO userSimpleDTO = modelMapper.map(user, UserSimpleDTO.class);
+
+            return new ResponseEntity<>(userSimpleDTO, HttpStatus.OK);
         } catch (Exception e) {
-            log.severe(e.getMessage());
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
-    }
-
-    @PostMapping("update/{id}")
-    public UserSimpleDTO updateUser(@RequestBody final UserUpdateDTO userDTO, @PathVariable String id) {
-        var user = jwtConvertor.getUserByJWT(id);
-        modelMapper.map(userDTO, user);
-        return modelMapper.map(
-                userService.createOrUpdate(user),
-                UserSimpleDTO.class);
-    }
-
-    @GetMapping("/{userId}/boards")
-    public ResponseEntity<List<BoardFullDTO>> getBoardsByUserId(@PathVariable Long userId) {
-        try {
-            return new ResponseEntity<>(userService.getBoardsByUserId(userId)
-                    .stream().map(board -> modelMapper.map(board,
-                            BoardFullDTO.class)).collect(Collectors.toList()),
-
-                    HttpStatus.OK);
-        } catch (CustomEntityNotFoundException e) {
-            log.severe(e.getMessage());
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
-    }
-
-    @GetMapping("{userId}/boardsByRole/{boardRoleId}")
-    public ResponseEntity<List<BoardFullDTO>> getBoardsByUserIdAndBoardRoleId(@PathVariable Long userId,
-                                                                              @PathVariable Long boardRoleId) {
-        try {
-            return new ResponseEntity<>(
-                    userService.getBoardsByUserIdAndBoardRoleId(userId, boardRoleId)
-                            .stream().map(board -> modelMapper.map(board,
-                            BoardFullDTO.class)).collect(Collectors.toList()),
-                    HttpStatus.OK);
-        } catch (CustomEntityNotFoundException e) {
-            log.severe(e.getMessage());
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
-    }
-
-    @GetMapping("/{userId}/comments")
-    public ResponseEntity<List<CommentDTO>> getCommentsByTicketId(@PathVariable Long userId) {
-        try {
-            return new ResponseEntity<>(
-                    userService.getAllCommentsByUserId(userId)
-                            .stream().map(comment ->
-                            modelMapper.map(comment, CommentDTO.class))
-                            .collect(Collectors.toList()),
-                    HttpStatus.OK);
-        } catch (CustomEntityNotFoundException e) {
             log.severe(e.getMessage());
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
